@@ -23,19 +23,19 @@
   },
   "homepage": "https://github.com/jaminitbot/Jam-Bot#readme",
   "dependencies": {
-    "@discordjs/builders": "^0.11.0",
-    "@discordjs/rest": "^0.2.0-canary.0",
-    "@prisma/client": "^3.7.0",
-    "@sentry/integrations": "^6.16.1",
-    "@sentry/node": "^6.16.1",
-    "@sentry/tracing": "^6.16.1",
-    "bson": "^4.6.0",
+    "@discordjs/builders": "^0.12.0",
+    "@discordjs/rest": "^0.3.0",
+    "@prisma/client": "^3.10.0",
+    "@sentry/integrations": "^6.18.0",
+    "@sentry/node": "^6.18.0",
+    "@sentry/tracing": "^6.18.0",
+    "bson": "^4.6.1",
     "date-fns": "^2.28.0",
-    "discord-api-types": "0.26.0",
-    "discord.js": "^13.5.0",
-    "dotenv": "^10.0.0",
-    "fastify": "^3.25.3",
-    "i18next": "^21.6.4",
+    "discord-api-types": "0.26.1",
+    "discord.js": "^13.6.0",
+    "dotenv": "^16.0.0",
+    "fastify": "^3.27.2",
+    "i18next": "^21.6.12",
     "i18next-fs-backend": "^1.1.4",
     "is-image-url": "^1.1.8",
     "is-number": "^7.0.0",
@@ -44,23 +44,23 @@
     "node-schedule": "^2.1.0",
     "prom-client": "^14.0.1",
     "tslib": "^2.3.1",
-    "undici": "^4.12.1",
-    "winston": "^3.3.3",
+    "undici": "^4.14.1",
+    "winston": "^3.6.0",
     "@jaminitbot/bot-utils": "^0.0.1"
   },
   "devDependencies": {
     "@types/i18next-fs-backend": "^1.1.2",
     "@types/is-number": "^7.0.2",
-    "@types/lodash": "^4.14.178",
+    "@types/lodash": "^4.14.179",
     "@types/ms": "^0.7.31",
-    "@types/node": "^17.0.7",
+    "@types/node": "^17.0.21",
     "@types/node-schedule": "^1.3.2",
-    "@typescript-eslint/eslint-plugin": "^5.8.1",
-    "@typescript-eslint/parser": "^5.8.1",
-    "eslint": "^8.6.0",
-    "prisma": "^3.7.0",
+    "@typescript-eslint/eslint-plugin": "^5.12.1",
+    "@typescript-eslint/parser": "^5.12.1",
+    "eslint": "^8.10.0",
+    "prisma": "^3.10.0",
     "standard-version": "^9.3.2",
-    "typescript": "^4.5.4"
+    "typescript": "^4.5.5"
   },
   "optionalDependencies": {
     "bufferutil": "^4.0.6",
