@@ -1,13 +1,11 @@
-import { AnyChannel, Guild, GuildMember, Message, Role } from 'discord.js'
-import { BotClient, Permission } from '../customDefinitions'
-import { getInvalidPermissionsMessage } from './messages'
-import { request } from 'undici'
-import { Logger } from 'winston'
-import i18next from 'i18next'
-import db from './db'
-import is_number = require('is-number')
-import { remove as removeFromArray } from 'lodash'
-import { BinaryLike, createHash } from 'crypto'
+import { GuildMember, Message } from "discord.js";
+import { BotClient, Permission } from "../customDefinitions";
+import { getInvalidPermissionsMessage } from "./messages";
+import { request } from "undici";
+import { Logger } from "winston";
+import i18next from "i18next";
+import db from "./db";
+import { remove as removeFromArray } from "lodash";
 
 /**
  * Checks permissions against a guild member
@@ -16,19 +14,19 @@ import { BinaryLike, createHash } from 'crypto'
  * @returns Boolean
  */
 export function checkPermissions(
-    member: GuildMember,
-    permissions: Array<Permission>
+  member: GuildMember,
+  permissions: Array<Permission>
 ): boolean {
-    let validPermission = true
-    if (permissions.includes('OWNER')) {
-        permissions = removeItemFromArray(permissions, 'OWNER')
-        if (!isBotOwner(member.id)) validPermission = false
-    }
-    if (permissions.length != 0) {
-        // @ts-expect-error
-        if (!member.permissions.has(permissions)) validPermission = false
-    }
-    return validPermission
+  let validPermission = true;
+  if (permissions.includes("OWNER")) {
+    permissions = removeItemFromArray(permissions, "OWNER");
+    if (!isBotOwner(member.id)) validPermission = false;
+  }
+  if (permissions.length != 0) {
+    // @ts-expect-error
+    if (!member.permissions.has(permissions)) validPermission = false;
+  }
+  return validPermission;
 }
 
 /**
@@ -37,22 +35,21 @@ export function checkPermissions(
  * @param stopCode Process exit code, default 0
  */
 export async function stopBot(
-    client: BotClient | null,
-    stopCode = 0
+  client: BotClient | null,
+  stopCode = 0
 ): Promise<void> {
-    try {
-        if (client) {
-            client.logger.warn(
-                'util: Received call to stop bot, stopping with code: ' +
-                stopCode
-            )
-            client.destroy()
-        }
-        await db.$disconnect()
-        process.exit(stopCode)
-    } catch {
-        process.exit()
+  try {
+    if (client) {
+      client.logger.warn(
+        "util: Received call to stop bot, stopping with code: " + stopCode
+      );
+      client.destroy();
     }
+    await db.$disconnect();
+    process.exit(stopCode);
+  } catch {
+    process.exit();
+  }
 }
 
 /**
@@ -62,9 +59,9 @@ export async function stopBot(
  * @returns Random number
  */
 export function randomInt(min: number, max: number): number {
-    min = Math.ceil(min)
-    max = Math.floor(max)
-    return Math.floor(Math.random() * (max - min + 1) + min) //The maximum is inclusive and the minimum is inclusive
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1) + min); //The maximum is inclusive and the minimum is inclusive
 }
 
 /**
@@ -72,114 +69,8 @@ export function randomInt(min: number, max: number): number {
  * @param message Initiating message
  */
 export function returnInvalidPermissionMessage(message: Message): void {
-    message.react('❌')
-    message.channel.send(getInvalidPermissionsMessage())
-}
-
-/**
- * Returns a user from a string, usually a ID or mention
- * @param guild Guild object
- * @param text Text to get the user from
- * @returns GuildMember
- */
-export async function getUserFromString(
-    guild: Guild,
-    text: unknown
-): Promise<GuildMember | null> {
-    try {
-        if (!text) return null
-        let stringText = String(text)
-        if (stringText.startsWith('<@') && stringText.endsWith('>')) {
-            // Mention
-            stringText = stringText.slice(2, -1)
-            if (stringText.startsWith('!')) {
-                stringText = stringText.slice(1)
-            }
-            if (stringText.startsWith('&')) {
-                // Role
-                return null
-            }
-            if (stringText.startsWith('<#')) {
-                // Channel
-                return null
-            }
-            return await guild.members.fetch(stringText)
-        } else if (is_number(text)) {
-            // Plain ID
-            return await guild.members.fetch(stringText)
-        }
-    } catch {
-        // eslint-disable-next-line no-empty
-    }
-    return null
-}
-
-/**
- * Returns a user from a string, usually a ID or mention
- * @param guild Guild object
- * @param text Text to get the user from
- * @returns GuildMember
- */
-export async function getRoleFromString(
-    guild: Guild,
-    text: unknown
-): Promise<Role | null> {
-    try {
-        if (!text) return null
-        let stringText = String(text)
-        if (stringText.startsWith('<@') && stringText.endsWith('>')) {
-            // Mention
-            stringText = stringText.slice(2, -1)
-            if (stringText.startsWith('<#')) {
-                // Channel
-                return null
-            }
-            if (stringText.startsWith('&')) {
-                // Role
-                stringText = stringText.slice(1)
-            }
-
-            return await guild.roles.fetch(stringText)
-        } else if (is_number(stringText)) {
-            // Plain ID
-            return await guild.roles.fetch(stringText)
-        }
-    } catch {
-        // eslint-disable-next-line no-empty
-    }
-    return null
-}
-
-/**
- * Returns a channel from a string of text, usually a ID or mention
- * @param guild Guild object
- * @param text Text to get channel from
- * @returns Channel
- */
-export async function getChannelFromString(
-    guild: Guild,
-    text: unknown
-): Promise<AnyChannel> {
-    try {
-        if (!text) return null
-        let stringText = String(text)
-        if (stringText.startsWith('<@')) {
-            // User or role
-            return null
-        }
-        if (stringText.startsWith('<#') && stringText.endsWith('>')) {
-            stringText = stringText.slice(2, -1)
-            return await guild.client.channels.fetch(stringText)
-        } else if (is_number(stringText)) {
-            return await guild.client.channels.fetch(stringText)
-        } else {
-            return guild.channels.cache.find(
-                (channel) => channel.name.toLowerCase() === stringText
-            )
-        }
-    } catch {
-        return null
-    }
+  message.react("❌");
+  message.channel.send(getInvalidPermissionsMessage());
 }
 
 /**
@@ -189,31 +80,29 @@ export async function getChannelFromString(
  * @returns string Uploaded paste location
  */
 export async function uploadToHasteBin(
-    logger: Logger,
-    dataToUpload: string
+  logger: Logger,
+  dataToUpload: string
 ): Promise<string | null> {
-    if (!dataToUpload) {
-        if (logger)
-            logger.error(
-                'hasteUploader: No content provided to upload, skipping...'
-            )
-    }
-    const hasteLocation = process.env.hasteBinHost ?? 'https://hastebin.com'
-    try {
-        const response = await request(hasteLocation + '/documents', {
-            method: 'POST',
-            body: dataToUpload,
-        })
-        if (response.statusCode != 200) return null
-        const responseData = await response.body.json()
-        if (responseData.key) return `${hasteLocation}/${responseData.key}`
-    } catch (err) {
-        if (logger)
-            logger.error(
-                'hasteUploader: Failed uploading to hastebin with error: ' + err
-            )
-    }
-    return null
+  if (!dataToUpload) {
+    if (logger)
+      logger.error("hasteUploader: No content provided to upload, skipping...");
+  }
+  const hasteLocation = process.env.hasteBinHost ?? "https://hastebin.com";
+  try {
+    const response = await request(hasteLocation + "/documents", {
+      method: "POST",
+      body: dataToUpload,
+    });
+    if (response.statusCode != 200) return null;
+    const responseData = await response.body.json();
+    if (responseData.key) return `${hasteLocation}/${responseData.key}`;
+  } catch (err) {
+    if (logger)
+      logger.error(
+        "hasteUploader: Failed uploading to hastebin with error: " + err
+      );
+  }
+  return null;
 }
 
 /**
@@ -222,51 +111,44 @@ export async function uploadToHasteBin(
  * @param value Value to remove
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function removeItemFromArray(array: Array<any>, value: unknown): Array<any> {
-    return removeFromArray(array, function (n: unknown) {
-        return value == n
-    })
+export function removeItemFromArray(
+  array: Array<any>,
+  value: unknown
+): Array<any> {
+  return removeFromArray(array, function (n: unknown) {
+    return value == n;
+  });
 }
 
-const owners = String(process.env.ownerId).split(',')
+const owners = String(process.env.ownerId).split(",");
 /**
  * Checks if a user ID is one of the bot owners
  * @param userId User ID to check
  * @returns Boolean
  */
 export function isBotOwner(userId: string) {
-    return owners.includes(userId)
+  return owners.includes(userId);
 }
 
-let thisLogger: Logger
+let thisLogger: Logger;
 export const saveLogger = (logger: Logger) => {
-    thisLogger = logger
-}
+  thisLogger = logger;
+};
 export const getLogger = () => {
-    return thisLogger
-}
+  return thisLogger;
+};
 
-/**
- * Capitalises the first letter of a sentence
- * @param string Input string
- * @returns string
- */
-export function capitaliseSentence(string: string) {
-    if (!string) return null
-    const str = String(string)
-    return str.charAt(0).toUpperCase() + str.slice(1)
-}
 /**
  * Converts a boolean value to a string representation
  * @param booleanToConvert Boolean to convert
  * @returns
  */
 export function booleanToHuman(booleanToConvert: boolean) {
-    if (booleanToConvert == true) {
-        return i18next.t('misc:ON')
-    } else {
-        return i18next.t('misc:OFF')
-    }
+  if (booleanToConvert == true) {
+    return i18next.t("misc:ON");
+  } else {
+    return i18next.t("misc:OFF");
+  }
 }
 
 /**
@@ -275,22 +157,9 @@ export function booleanToHuman(booleanToConvert: boolean) {
  * @returns  Promise
  */
 export async function delay(time: number) {
-    return new Promise((resolve) => setTimeout(resolve, time))
+  return new Promise((resolve) => setTimeout(resolve, time));
 }
-
-/**
- * Hashes data using a specific hash type
- * @param data data to hash
- * @param hashType type of hash to use
- * @returns hash
- */
-export function hash(data: BinaryLike, hashType: string) {
-    const hash = createHash(hashType)
-        .update(data)
-        .digest('hex')
-    return hash
-}
-
+// prettier-ignore
 const emojis = [
     '😄', '😃', '😀', '😊', '☺', '😉', '😍', '😘', '😚', '😗', '😙', '😜', '😝', '😛', '😳', '😁', '😔', '😌', '😒', '😞', '😣', '😢', '😂', '😭', '😪', '😥', '😰', '😅', '😓', '😩', '😫', '😨', '😱', '😠', '😡', '😤', '😖', '😆', '😋', '😷', '😎', '😴', '😵', '😲', '😟', '😦', '😧', '😈', '👿', '😮', '😬', '😐', '😕', '😯', '😶', '😇', '😏', '😑', '👲', '👳', '👮', '👷', '💂', '👶', '👦', '👧', '👨', '👩', '👴', '👵', '👱', '👼', '👸', '😺', '😸', '😻', '😽', '😼', '🙀', '😿', '😹', '😾', '👹', '👺', '🙈', '🙉', '🙊', '💀', '👽', '💩', '🔥', '✨', '🌟', '💫', '💥', '💢', '💦', '💧', '💤', '💨', '👂', '👀', '👃', '👅', '👄', '👍', '👎', '👌', '👊', '✊', '✌', '👋', '✋', '👐', '👆', '👇', '👉', '👈', '🙌', '🙏', '☝', '👏', '💪', '🚶', '🏃', '💃', '👫', '👪', '👬', '👭', '💏', '💑', '👯', '🙆', '🙅', '💁', '🙋', '💆', '💇', '💅', '👰', '🙎', '🙍', '🙇', '🎩', '👑', '👒', '👟', '👞', '👡', '👠', '👢', '👕', '👔', '👚', '👗', '🎽', '👖', '👘', '👙', '💼', '👜', '👝', '👛', '👓', '🎀', '🌂', '💄', '💛', '💙', '💜', '💚', '❤', '💔', '💗', '💓', '💕', '💖', '💞', '💘', '💌', '💋', '💍', '💎', '👤', '👥', '💬', '👣', '💭', '🐶', '🐺', '🐱', '🐭', '🐹', '🐰', '🐸', '🐯', '🐨', '🐻', '🐷', '🐽', '🐮', '🐗', '🐵', '🐒', '🐴', '🐑', '🐘', '🐼', '🐧', '🐦', '🐤', '🐥', '🐣', '🐔', '🐍', '🐢', '🐛', '🐝', '🐜', '🐞', '🐌', '🐙', '🐚', '🐠', '🐟', '🐬', '🐳', '🐋', '🐄', '🐏', '🐀', '🐃', '🐅', '🐇', '🐉', '🐎', '🐐', '🐓', '🐕', '🐖', '🐁', '🐂', '🐲', '🐡', '🐊', '🐫', '🐪', '🐆', '🐈', '🐩', '🐾', '💐', '🌸', '🌷', '🍀', '🌹', '🌻', '🌺', '🍁', '🍃', '🍂', '🌿', '🌾', '🍄', '🌵', '🌴', '🌲', '🌳', '🌰', '🌱', '🌼', '🌐', '🌞', '🌝', '🌚', '🌑', '🌒', '🌓', '🌔', '🌕', '🌖', '🌗', '🌘', '🌜', '🌛', '🌙', '🌍', '🌎', '🌏', '🌋', '🌌', '🌠', '⭐', '☀', '⛅', '☁', '⚡', '☔', '❄', '⛄', '🌀', '🌁', '🌈', '🌊', '🎍', '💝', '🎎', '🎒', '🎓', '🎏', '🎆', '🎇', '🎐', '🎑', '🎃', '👻', '🎅', '🎄', '🎁', '🎋', '🎉', '🎊', '🎈', '🎌', '🔮', '🎥', '📷', '📹', '📼', '💿', '📀', '💽', '💾', '💻', '📱', '☎', '📞', '📟', '📠', '📡', '📺', '📻', '🔊', '🔉', '🔈', '🔇', '🔔', '🔕', '📢', '📣', '⏳', '⌛', '⏰', '⌚', '🔓', '🔒', '🔏', '🔐', '🔑', '🔎', '💡', '🔦', '🔆', '🔅', '🔌', '🔋', '🔍', '🛁', '🛀', '🚿', '🚽', '🔧', '🔩', '🔨', '🚪', '🚬', '💣', '🔫', '🔪', '💊', '💉', '💰', '💴', '💵', '💷', '💶', '💳', '💸', '📲', '📧', '📥', '📤', '✉', '📩', '📨', '📯', '📫', '📪', '📬', '📭', '📮', '📦', '📝', '📄', '📃', '📑', '📊', '📈', '📉', '📜', '📋', '📅', '📆', '📇', '📁', '📂', '✂', '📌', '📎', '✒', '✏', '📏', '📐', '📕', '📗', '📘', '📙', '📓', '📔', '📒', '📚', '📖', '🔖', '📛', '🔬', '🔭', '📰', '🎨', '🎬', '🎤', '🎧', '🎼', '🎵', '🎶', '🎹', '🎻', '🎺', '🎷', '🎸', '👾', '🎮', '🃏', '🎴', '🀄', '🎲', '🎯', '🏈', '🏀', '⚽', '⚾', '🎾', '🎱', '🏉', '🎳', '⛳', '🚵', '🚴', '🏁', '🏇', '🏆', '🎿', '🏂', '🏊', '🏄', '🎣', '☕', '🍵', '🍶', '🍼', '🍺', '🍻', '🍸', '🍹', '🍷', '🍴', '🍕', '🍔', '🍟', '🍗', '🍖', '🍝', '🍛', '🍤', '🍱', '🍣', '🍥', '🍙', '🍘', '🍚', '🍜', '🍲', '🍢', '🍡', '🍳', '🍞', '🍩', '🍮', '🍦', '🍨', '🍧', '🎂', '🍰', '🍪', '🍫', '🍬', '🍭', '🍯', '🍎', '🍏', '🍊', '🍋', '🍒', '🍇', '🍉', '🍓', '🍑', '🍈', '🍌', '🍐', '🍍', '🍠', '🍆', '🍅', '🌽', '🏠', '🏡', '🏫', '🏢', '🏣', '🏥', '🏦', '🏪', '🏩', '🏨', '💒', '⛪', '🏬', '🏤', '🌇', '🌆', '🏯', '🏰', '⛺', '🏭', '🗼', '🗾', '🗻', '🌄', '🌅', '🌃', '🗽', '🌉', '🎠', '🎡', '⛲', '🎢', '🚢', '⛵', '🚤', '🚣', '⚓', '🚀', '✈', '💺', '🚁', '🚂', '🚊', '🚉', '🚞', '🚆', '🚄', '🚅', '🚈', '🚇', '🚝', '🚋', '🚃', '🚎', '🚌', '🚍', '🚙', '🚘', '🚗', '🚕', '🚖', '🚛', '🚚', '🚨', '🚓', '🚔', '🚒', '🚑', '🚐', '🚲', '🚡', '🚟', '🚠', '🚜', '💈', '🚏', '🎫', '🚦', '🚥', '⚠', '🚧', '🔰', '⛽', '🏮', '🎰', '♨', '🗿', '🎪', '🎭', '📍', '🚩', '⬆', '⬇', '⬅', '➡', '🔠', '🔡', '🔤', '↗', '↖', '↘', '↙', '↔', '↕', '🔄', '◀', '▶', '🔼', '🔽', '↩', '↪', 'ℹ', '⏪', '⏩', '⏫', '⏬', '⤵', '⤴', '🆗', '🔀', '🔁', '🔂', '🆕', '🆙', '🆒', '🆓', '🆖', '📶', '🎦', '🈁', '🈯', '🈳', '🈵', '🈴', '🈲', '🉐', '🈹', '🈺', '🈶', '🈚', '🚻', '🚹', '🚺', '🚼', '🚾', '🚰', '🚮', '🅿', '♿', '🚭', '🈷', '🈸', '🈂', 'Ⓜ', '🛂', '🛄', '🛅', '🛃', '🉑', '㊙', '㊗', '🆑', '🆘', '🆔', '🚫', '🔞', '📵', '🚯', '🚱', '🚳', '🚷', '🚸', '⛔', '✳', '❇', '❎', '✅', '✴', '💟', '🆚', '📳', '📴', '🅰', '🅱', '🆎', '🅾', '💠', '➿', '♻', '♈', '♉', '♊', '♋', '♌', '♍', '♎', '♏', '♐', '♑', '♒', '♓', '⛎', '🔯', '🏧', '💹', '💲', '💱', '©', '®', '™', '〽', '〰', '🔝', '🔚', '🔙', '🔛', '🔜', '❌', '⭕', '❗', '❓', '❕', '❔', '🔃', '🕛', '🕧', '🕐', '🕜', '🕑', '🕝', '🕒', '🕞', '🕓', '🕟', '🕔', '🕠', '🕕', '🕖', '🕗', '🕘', '🕙', '🕚', '🕡', '🕢', '🕣', '🕤', '🕥', '🕦', '✖', '➕', '➖', '➗', '♠', '♥', '♣', '♦', '💮', '💯', '✔', '☑', '🔘', '🔗', '➰', '🔱', '🔲', '🔳', '◼', '◻', '◾', '◽', '▪', '▫', '🔺', '⬜', '⬛', '⚫', '⚪', '🔴', '🔵', '🔻', '🔶', '🔷', '🔸', '🔹'
 ]
@@ -300,19 +169,11 @@ const emojis = [
  * @returns Emoji
  */
 export function randomEmoji() {
-    return emojis[Math.floor(Math.random() * emojis.length)]
+  return emojis[Math.floor(Math.random() * emojis.length)];
 }
 
-/**
- * Returns a random hex code
- * @returns Hex Code
- */
-export function randomHexCode() {
-    return '#' + Math.floor(Math.random() * 16777215).toString(16)
-}
-
-import { GLOBAL_RATELIMIT_DURATION } from '../consts'
-const rateLimits = new Map()
+import { GLOBAL_RATELIMIT_DURATION } from "../consts";
+const rateLimits = new Map();
 
 /**
  * Checks whether a user should be rate limited
@@ -321,11 +182,15 @@ const rateLimits = new Map()
  * @param userId User ID to check
  * @returns Boolean
  */
-export function checkRateLimit(commandName: string, commandLimit: number | undefined, userId: string): boolean {
-    if (!commandLimit) commandLimit = GLOBAL_RATELIMIT_DURATION
-    const rateLimit = rateLimits.get(`${commandName}-${userId}`) ?? 0
-    if (Date.now() < (commandLimit * 1000) + rateLimit) return true
-    return false
+export function checkRateLimit(
+  commandName: string,
+  commandLimit: number | undefined,
+  userId: string
+): boolean {
+  if (!commandLimit) commandLimit = GLOBAL_RATELIMIT_DURATION;
+  const rateLimit = rateLimits.get(`${commandName}-${userId}`) ?? 0;
+  if (Date.now() < commandLimit * 1000 + rateLimit) return true;
+  return false;
 }
 
 /**
@@ -334,7 +199,7 @@ export function checkRateLimit(commandName: string, commandLimit: number | undef
  * @param userId User ID to rate limit
  */
 export function setRateLimit(commandName: string, userId: string) {
-    rateLimits.set(`${commandName}-${userId}`, Date.now())
+  rateLimits.set(`${commandName}-${userId}`, Date.now());
 }
 
 /**
@@ -344,36 +209,43 @@ export function setRateLimit(commandName: string, userId: string) {
  * @param userId User ID to check
  * @returns Time remaining in MS
  */
-export function getRateLimitRemaining(commandName: string, commandLimit: number | undefined, userId: string) {
-    if (!commandLimit) commandLimit = GLOBAL_RATELIMIT_DURATION
-    const rateLimit = rateLimits.get(`${commandName}-${userId}`)
-    if (!rateLimit) return 0
-    return commandLimit * 1000 - (Date.now() - rateLimit)
+export function getRateLimitRemaining(
+  commandName: string,
+  commandLimit: number | undefined,
+  userId: string
+) {
+  if (!commandLimit) commandLimit = GLOBAL_RATELIMIT_DURATION;
+  const rateLimit = rateLimits.get(`${commandName}-${userId}`);
+  if (!rateLimit) return 0;
+  return commandLimit * 1000 - (Date.now() - rateLimit);
 }
 
 export function warnForOmittedEnvs(logger: Logger) {
-    if (!process.env.ownerId || !process.env.ownerName) {
-        logger.warn('Owner information not provided in .env')
-    }
-    if (!process.env.guildLogChannel || !process.env.dmChannel) {
-        logger.warn('Guild log channel(s) not provided in .env')
-    }
-    if (!process.env.devServerId) {
-        logger.warn('Dev server ID not provided in .env')
-    }
-    if (!process.env.errorLogWebhookUrl) {
-        logger.warn('Error webhook not provided in .env')
-    }
-    if (!process.env.sentryUrl) {
-        logger.warn('Sentry url not provided in .env')
-    }
-    if (!process.env.twitchNotificationsChannel || !process.env.twitchNotificationsUsername) {
-        logger.warn('Twitch channel information not provided in .env')
-    }
-    if (!process.env.twitchApiSecret || !process.env.twitchApiClientId) {
-        logger.warn('Twitch auth credentials not provided in .env')
-    }
-    if (!process.env.bingImageSearchKey || !process.env.pexelsApiKey) {
-        logger.warn('Image auth credentials not specified in .env')
-    }
+  if (!process.env.ownerId || !process.env.ownerName) {
+    logger.warn("Owner information not provided in .env");
+  }
+  if (!process.env.guildLogChannel || !process.env.dmChannel) {
+    logger.warn("Guild log channel(s) not provided in .env");
+  }
+  if (!process.env.devServerId) {
+    logger.warn("Dev server ID not provided in .env");
+  }
+  if (!process.env.errorLogWebhookUrl) {
+    logger.warn("Error webhook not provided in .env");
+  }
+  if (!process.env.sentryUrl) {
+    logger.warn("Sentry url not provided in .env");
+  }
+  if (
+    !process.env.twitchNotificationsChannel ||
+    !process.env.twitchNotificationsUsername
+  ) {
+    logger.warn("Twitch channel information not provided in .env");
+  }
+  if (!process.env.twitchApiSecret || !process.env.twitchApiClientId) {
+    logger.warn("Twitch auth credentials not provided in .env");
+  }
+  if (!process.env.bingImageSearchKey || !process.env.pexelsApiKey) {
+    logger.warn("Image auth credentials not specified in .env");
+  }
 }
